refactor api server, add datatag api
use different server process api and web
rename tags field
merge json read/write

+++ src/parser/neu_json_login.h
@@ -33,7 +33,7 @@ struct neu_parse_login_req {
 struct neu_parse_login_res {
     enum neu_parse_function function;
     char *                  uuid;
-    int                     error;
+    int64_t                 error;
     char *                  token;
 };
